fix: wrong request structure & loading container
