+++ src/points-api-service.js
@@ -39,7 +39,7 @@ export default class PointsApiService extends ApiService {
       'id': point.id,
       'date_from': point.dateFrom instanceof Date ? point.dateFrom.toISOString() : null,
       'date_to': point.dateTo instanceof Date ? point.dateTo.toISOString() : null,
-      'base_price': point.basePrice,
+      'base_price': parseInt(point.basePrice, 10),
       'is_favorite': point.isFavorite
     };
